added explicit type conversion

+++ SMBActions.m
@@ -1,13 +1,13 @@
 /* ######################################################################
 * File Name: SMBActions.m
-* Project: ssp
+* Project: SSP
 * Version: 18.10
 * Creation Date: 10.10.2018
-* Created By Sebastian Malkusch
-* <malkusch@chemie.uni-frankfurt.de>
-* Goethe University of Frankfurt
-* Physical and Theoretical Chemistry
-* Single Molecule Biophysics
+* Created By: Sebastian Malkusch
+* Contact: <malkusch@chemie.uni-frankfurt.de>
+* Company: Goethe University of Frankfurt
+* Institute: Physical and Theoretical Chemistry
+* Department: Single Molecule Biophysics
 *
 * License
 * Copyright (C) 2018  Sebastian Malkusch
@@ -41,7 +41,7 @@
 
 -(void) runActions
 {
-	srand (time(NULL));
+	srand ((unsigned int)time(NULL));
 }
 
 -(void) dealloc
